+++ src/contexts/index.ts
@@ -1,1 +1,2 @@
 export { AppStorageProvider, useAppStorageContext } from "./AppStorageProvider";
+export { SettingsProvider, useSettingsContext } from "./SettingsProvider";